Refactor file structure of GeoJSON
